-Wsign-conversion and ssize_t fixes
Additional changes:
- Delete unused dllopen code.
- Removed STR_INT from the strgrpmatch().
- Fix contra least surprise and unused ssizeof()
  macro; use ssizeof() macro.
- Fixed erroneous tautology in AST iconv().

This is still not close to done, yet all of this needs to
become a submittable patch series, somehow...

+++ src/cmd/ksh93/sh/io.c
@@ -2642,7 +2642,7 @@ Sfio_t	*sh_fd2sfio(int fd)
 			flags |= SFIO_READ;
 		if(status&IOWRITE)
 			flags |= SFIO_WRITE;
-		sp = sfnew(NULL, NULL, -1, fd,flags);
+		sp = sfnew(NULL, NULL, (size_t)-1, fd,flags);
 		sh.sftable[fd] = sp;
 	}
 	return sp;
