+++ src/lib/libast/string/swapget.c
@@ -32,15 +32,15 @@
  */
 
 intmax_t
-swapget(int op, const void* b, int n)
+swapget(int op, const void* b, size_t n)
 {
 	unsigned char*	p;
 	unsigned char*	d;
 	intmax_t		v;
 	unsigned char		tmp[sizeof(intmax_t)];
 
-	if (n > (int)sizeof(intmax_t))
-		n = (int)sizeof(intmax_t);
+	if (n > sizeof(intmax_t))
+		n = sizeof(intmax_t);
 	if (op) swapmem(op, b, d = tmp, n);
 	else d = (unsigned char*)b;
 	p = d + n;
